+++ src/components/modals/index.js
@@ -29,6 +29,7 @@ import WalletType from 'components/modals/device/WalletType';
 import Nem from 'components/modals/external/Nem';
 import Cardano from 'components/modals/external/Cardano';
 import Stellar from 'components/modals/external/Stellar';
+import Tezos from 'components/modals/external/Tezos';
 
 import QrModal from 'components/modals/QrModal';
 
@@ -168,7 +169,9 @@ const getExternalContextModal = (props: Props) => {
         case 'xlm':
             return <Stellar onCancel={modalActions.onCancel} />;
         case 'ada':
-            return <Cardano onCancel={modalActions.onCancel} />;
+            return (<Cardano onCancel={modalActions.onCancel} />);
+        case 'xtz':
+            return (<Tezos onCancel={modalActions.onCancel} />);
         default:
             return null;
     }