+++ src/views/Wallet/views/Account/Send/ripple/components/AdvancedForm/index.js
@@ -82,16 +82,16 @@ const getFeeInputState = (feeErrors: string, feeWarnings: string): string => {
     return state;
 };
 
-// const getDestinationTagInputState = (errors: string, warnings: string): string => {
-//     let state = '';
-//     if (warnings && !errors) {
-//         state = 'warning';
-//     }
-//     if (errors) {
-//         state = 'error';
-//     }
-//     return state;
-// };
+const getDestinationTagInputState = (errors: string, warnings: string): string => {
+    let state = '';
+    if (warnings && !errors) {
+        state = 'warning';
+    }
+    if (errors) {
+        state = 'error';
+    }
+    return state;
+};
 
 const Left = styled.div`
     display: flex;
@@ -107,11 +107,11 @@ const AdvancedForm = (props: Props) => {
         warnings,
         infos,
         fee,
-        // destinationTag,
+        destinationTag,
     } = props.sendForm;
     const {
         onFeeChange,
-        // onDestinationTagChange,
+        onDestinationTagChange,
     } = props.sendFormActions;
 
     return (
@@ -150,7 +150,7 @@ const AdvancedForm = (props: Props) => {
                 />
             </InputRow>
 
-            {/* <InputRow>
+            <InputRow>
                 <StyledInput
                     state={getDestinationTagInputState(errors.destinationTag, warnings.destinationTag)}
                     autoComplete="off"
@@ -180,7 +180,7 @@ const AdvancedForm = (props: Props) => {
                     value={destinationTag}
                     onChange={event => onDestinationTagChange(event.target.value)}
                 />
-            </InputRow> */}
+            </InputRow>
 
             <AdvancedSettingsSendButtonWrapper>{props.children}</AdvancedSettingsSendButtonWrapper>
         </AdvancedSettingsWrapper>